Rename files uses JSX syntax from .js to .jsx
